made chages for UI

+++ app/src/main/java/com/vishalmusale/weather/network/model/WeatherDto.kt
@@ -12,7 +12,7 @@ data class WeatherDto (
     @SerializedName("visibility" ) var visibility : Int?               = null,
     @SerializedName("wind"       ) var wind       : Wind?              = Wind(),
     @SerializedName("clouds"     ) var clouds     : Clouds?            = Clouds(),
-    @SerializedName("dt"         ) var dt         : Int?               = null,
+    @SerializedName("dt"         ) var dt         : Long?               = null,
     @SerializedName("sys"        ) var sys        : Sys?               = Sys(),
     @SerializedName("timezone"   ) var timezone   : Int?               = null,
     @SerializedName("id"         ) var id         : Int?               = null,
